Move common functionality to ABC
